Edited Logos, Added grid and cards, CSS styling
Changed the Education section to a responsive grid.
Changed the Experience section to a grid with cards of equal height.
Edited logos for a better fit with Experience cards, and a squarer shaped Github logo in the footer.
Changed CSS code for the webpage background color and gave a different font color for the cards.